fix: 앰플리파이 레디스 설정
