+++ src/shared/libs/handler/token.ts
@@ -1,5 +1,5 @@
-import { redlock, RedlockLock } from '@/shared/libs/redis/lock.server';
-import { redis } from '@/shared/libs/redis/redis.server';
+import { redlock, RedlockLock } from '@/shared/libs/redis/lock';
+import { redis } from '@/shared/libs/redis/redis';
 import { refreshAccessToken } from './refreshAccessToken';
 
 export interface RefreshTokenResult {
