--- src/shared/libs/redis/redis.server.ts
@@ -1,7 +1,0 @@
-import Redis from 'ioredis';
-
-export const redis = new Redis({
-  host: process.env.REDIS_HOST,
-  port: Number(process.env.REDIS_PORT),
-  password: process.env.REDIS_PASSWORD,
-});
